+++ dartbrcode/lib/dartbrcode.dart
@@ -1,3 +1,4 @@
+import 'dart:convert';
 import 'dart:ffi' as ffi;
 import 'dart:ffi';
 import 'dart:io';
@@ -7,13 +8,25 @@ typedef NativeRustUt8ToUt8 = ffi.Pointer<Utf8> Function(ffi.Pointer<Utf8>);
 
 final osSpecificFile = Platform.isMacOS
     ? "libbrcode.dylib"
-    : Platform.isIOS ? "libbrcode.a" : "libbrcode.so";
+    : Platform.isIOS
+        ? "libbrcode.a"
+        : "libbrcode.so";
 final ffi.DynamicLibrary dl = ffi.DynamicLibrary.open(osSpecificFile);
 
 String jsonToBrcode(String json) {
   final json_to_brcode = dl
       .lookupFunction<NativeRustUt8ToUt8, NativeRustUt8ToUt8>("json_to_brcode");
   final Pointer<Utf8> utf8_brcode = Utf8.toUtf8(json).cast();
+  final utf8_from_json = json_to_brcode(utf8_brcode);
+
+  return Utf8.fromUtf8(utf8_from_json).toString();
+}
+
+String jsonMapToBrcode(Map<String, dynamic> json) {
+  final jsonStr = jsonEncode(json);
+  final json_to_brcode = dl
+      .lookupFunction<NativeRustUt8ToUt8, NativeRustUt8ToUt8>("json_to_brcode");
+  final Pointer<Utf8> utf8_brcode = Utf8.toUtf8(jsonStr).cast();
   final utf8_from_json = json_to_brcode(utf8_brcode);
 
   return Utf8.fromUtf8(utf8_from_json).toString();
@@ -27,6 +40,16 @@ String brcodeToJson(String code) {
   final utf8_to_json = json_from_brcode(utf8_brcode);
 
   return Utf8.fromUtf8(utf8_to_json).toString();
+}
+
+Map<String, dynamic> brcodeToJsonMap(String code) {
+  final json_from_brcode =
+      dl.lookupFunction<NativeRustUt8ToUt8, NativeRustUt8ToUt8>(
+          "json_from_brcode");
+  final Pointer<Utf8> utf8_brcode = Utf8.toUtf8(code).cast();
+  final utf8_to_json = json_from_brcode(utf8_brcode);
+
+  return jsonDecode(Utf8.fromUtf8(utf8_to_json).toString());
 }
 
 String crc16Ccitt(String code) {