+++ dartbrcode/CHANGELOG.md
@@ -1,6 +1,10 @@
-# 0.2.0
+## 0.3.0
+- `jsonMapToBrcode` - function to encode a Json `Map<String, dynamic>` as a brcode string.
+- `brcodeToJsonMap` - function to parse a brcode string as a Json `Map<String, dynamic>`.
+
+## 0.2.0
 - `crc16Ccitt` - returns checksum string of a string message.
 
-# 0.1.0
+## 0.1.0
 - `jsonToBrcode` - function to encode a Json as a brcode string.
 - `brcodeToJson` - function to parse a brcode string as a Json.